объединила фрагменты
