+++ app/schemas/ru.alinadorozhkina.musicapp.mvp.model.entity.room.db.DataBase/5.json
@@ -1,7 +1,7 @@
 {
   "formatVersion": 1,
   "database": {
-    "version": 4,
+    "version": 5,
     "identityHash": "4b85479885edbc06a1fcacc24eb21e56",
     "entities": [
       {
