+++ app/src/main/java/ru/alinadorozhkina/musicapp/mvp/model/entity/room/db/DataBase.kt
@@ -16,7 +16,7 @@ import ru.alinadorozhkina.musicapp.mvp.model.entity.room.dao.ImageDao
         RoomArtist::class,
         RoomCachedImage::class
     ],
-    version = 4,
+    version = 5,
     exportSchema = true
 )
 abstract class DataBase : RoomDatabase() {
@@ -26,7 +26,7 @@ abstract class DataBase : RoomDatabase() {
     abstract val imageDao: ImageDao
 
     companion object {
-        const val DB_NAME = "data_base4.db"
+        const val DB_NAME = "data_base5.db"
         private var instance: DataBase? = null
         fun create(context: Context) {
             if (instance == null) {
