--- app/src/main/java/ru/alinadorozhkina/musicapp/mvp/model/view/TrackListView.kt
@@ -1,11 +1,0 @@
-package ru.alinadorozhkina.musicapp.mvp.model.view
-
-import moxy.MvpView
-import moxy.viewstate.strategy.AddToEndSingleStrategy
-import moxy.viewstate.strategy.StateStrategyType
-
-@StateStrategyType(AddToEndSingleStrategy::class)
-interface TrackLisView : MvpView {
-    fun init()
-    fun updateList()
-}
