+++ app/src/main/java/ru/alinadorozhkina/musicapp/mvp/navigation/IScreens.kt
@@ -6,6 +6,5 @@ import ru.alinadorozhkina.musicapp.mvp.model.entity.Artist
 interface IScreens {
     fun topList(): Screen
     fun artist(artist: Artist): Screen
-    fun artistTracks(artist: Artist): Screen
     fun settings(): Screen
 }
