+++ app/src/main/java/ru/alinadorozhkina/musicapp/ui/activities/MainActivity.kt
@@ -45,7 +45,6 @@ class MainActivity : MvpAppCompatActivity(), MainView {
         setContentView(ui?.root)
     }
 
-
     override fun onResumeFragments() {
         super.onResumeFragments()
         navigatorHolder.setNavigator(navigator)
