+++ app/src/main/java/ru/alinadorozhkina/musicapp/ui/navigation/AndroidScreens.kt
@@ -4,16 +4,12 @@ import com.github.terrakok.cicerone.Screen
 import com.github.terrakok.cicerone.androidx.FragmentScreen
 import ru.alinadorozhkina.musicapp.mvp.model.entity.Artist
 import ru.alinadorozhkina.musicapp.mvp.navigation.IScreens
-import ru.alinadorozhkina.musicapp.ui.fragments.ArtistFragment
-import ru.alinadorozhkina.musicapp.ui.fragments.ArtistTracksFragment
-import ru.alinadorozhkina.musicapp.ui.fragments.SettingsFragment
-import ru.alinadorozhkina.musicapp.ui.fragments.TopTracksFragment
+import ru.alinadorozhkina.musicapp.ui.fragments.*
 
 class AndroidScreens : IScreens {
 
     override fun topList() = FragmentScreen { TopTracksFragment.newInstance() }
-    override fun artist(artist: Artist) = FragmentScreen { ArtistFragment.newInstance(artist) }
-    override fun artistTracks(artist: Artist): Screen = FragmentScreen { ArtistTracksFragment.newInstance(artist) }
+    override fun artist(artist: Artist) = FragmentScreen { ArtistInfoFragment.newInstance(artist) }
     override fun settings(): Screen = FragmentScreen { SettingsFragment.newInstance()}
 
 }
