+++ app/src/main/res/layout/bottom_sheet.xml
@@ -1,44 +1,53 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
+    android:layout_height="wrap_content"
+    android:id="@+id/bottom_sheet"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     app:behavior_hideable="false"
     app:behavior_peekHeight="100dp"
     android:background="@color/gold"
     app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior">
 
-<TextView
-        android:id="@+id/name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"/>
 
-    <TextView
-        android:id="@+id/nb_album"
-        android:layout_width="wrap_content"
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:orientation="horizontal"
+        android:background="?attr/colorSecondaryVariant"
+        android:layout_marginTop="@dimen/margin_padding_size_small"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/name"/>
+        app:layout_constraintTop_toTopOf="parent">
 
-    <TextView
-        android:id="@+id/fans"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/nb_album"/>
 
-    <TextView
-        android:id="@+id/radio"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/fans"/>
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/button_favourites"
+            style="@style/MyIconButtonStyle"
+            app:icon="@drawable/ic_baseline_favorite_border_24"
+            />
+        <SeekBar
+            android:id="@+id/seekbar"
+            android:layout_margin="16dp"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center"
+            android:layout_weight="1"/>
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/button_play"
+            style="@style/MyIconButtonStyle"
+
+            app:icon="@android:drawable/ic_media_play" />
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/button_stop"
+            style="@style/MyIconButtonStyle"
+            app:icon="@android:drawable/ic_media_pause"
+            android:layout_marginEnd="@dimen/margin_padding_size_small"/>
+
+
+    </LinearLayout>
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
