--- app/src/main/res/layout/fragment_artist.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    style="@style/viewBackgroundStyle"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.fragments.ArtistFragment">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/tv_artist_name"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="@dimen/margin_padding_size_medium"
-        style="@style/BaseText.main_text" />
-
-    <ImageView
-        android:id="@+id/artist_picture"
-        android:layout_width="@dimen/poster_size"
-        android:layout_height="@dimen/poster_size"
-        android:layout_gravity="center"/>
-
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/btn_show_trackList"
-        style="@style/MyButtonStyle"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Показать трек лист"
-        app:icon="@drawable/ic_baseline_library_music_24"
-        android:layout_margin="@dimen/margin_padding_size_medium"/>
-
-    </LinearLayout>
-
-    <include layout="@layout/bottom_sheet" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
