--- app/src/main/res/layout/fragment_artist_tracks.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    style="@style/viewBackgroundStyle"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.fragments.ArtistTracksFragment">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_artist_tracks"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:paddingTop="@dimen/margin_padding_size_medium"
-        android:layout_margin="@dimen/margin_padding_size_small"
-        android:clipToPadding="false"
-        android:paddingBottom="@dimen/margin_padding_size_medium" />
-
-</FrameLayout>
